created financialEventControllers.js
